refactor(clients): split protocols.py into subpackage
Split monolithic clients/protocols.py into clients/protocols/ subpackage
to improve modularity and architectural clarity.

Changes:
- Create clients/protocols/ subpackage with 6 modules:
  - backend.py: BackendProtocol
  - github.py: GitHubClientProtocol and narrow port protocols
  - flow.py: FlowReader (migrated from services/flow_reader.py)
  - git.py: GitPathProtocol (extracted from services/git_path_client.py)
  - pr.py: BaseResolver (extracted from services/pr_create_usecase.py)
  - __init__.py: Re-exports all public symbols for backward compatibility
- Delete clients/protocols.py (replaced by subpackage)
- Update services/flow_reader.py to re-export shim
- Update services/git_path_client.py to import GitPathProtocol from new location
- Update services/pr_create_usecase.py to import BaseResolver from new location

Backward compatibility:
- All existing imports continue to work via __init__.py re-exports
- services/ modules become thin shims delegating to clients/protocols/
- Two different GitClientProtocol definitions remain separate:
  - clients/git_client.py: GitClientProtocol for git operations
  - services/git_path_client.py: GitClientProtocol alias for path operations

Verification:
- All 2768 tests pass
- mypy: Success, no issues in 395 source files
- ruff: All checks passed

Closes #2110
