+++ src/vibe3/clients/protocols/git.py
@@ -1,0 +1,16 @@
+"""GitPathProtocol — Protocol for Git path-related operations."""
+
+from pathlib import Path
+from typing import Protocol
+
+
+class GitPathProtocol(Protocol):
+    """Protocol for Git path-related operations."""
+
+    def get_git_common_dir(self) -> str: ...
+    def get_worktree_root(self) -> str: ...
+    def find_worktree_path_for_branch(self, branch: str) -> Path | None: ...
+    def get_current_branch(self) -> str: ...
+
+
+__all__ = ["GitPathProtocol"]
