+++ src/vibe3/services/git_path_client.py
@@ -1,16 +1,8 @@
 """Git path operations protocol and wrappers."""
 
 from pathlib import Path
-from typing import Protocol
 
-
-class GitPathProtocol(Protocol):
-    """Protocol for Git path-related operations."""
-
-    def get_git_common_dir(self) -> str: ...
-    def get_worktree_root(self) -> str: ...
-    def find_worktree_path_for_branch(self, branch: str) -> Path | None: ...
-    def get_current_branch(self) -> str: ...
+from vibe3.clients.protocols.git import GitPathProtocol
 
 
 def _get_git_client(git_client: GitPathProtocol | None = None) -> GitPathProtocol:
