+++ src/vibe3/services/pr_create_usecase.py
@@ -2,7 +2,6 @@
 
 import os
 from dataclasses import dataclass
-from typing import Protocol
 
 from loguru import logger
 from rich.console import Console
@@ -10,6 +9,7 @@ from rich.prompt import Prompt
 
 from vibe3.clients.ai_client import AIClient
 from vibe3.clients.ai_suggestion_client import AISuggestionClient
+from vibe3.clients.protocols.pr import BaseResolver
 from vibe3.config.settings import VibeConfig
 from vibe3.prompts.template_loader import resolve_prompts_path
 from vibe3.services.base_resolution_usecase import BaseResolutionUsecase
@@ -25,13 +25,6 @@ class PRCreateResult:
     body: str
     base_branch: str
     actor: str
-
-
-class BaseResolver(Protocol):
-    """Protocol for base branch resolution."""
-
-    def resolve_pr_create_base(self, requested_base: str | None) -> str: ...
-    def collect_branch_material(self, base_branch: str, branch: str) -> object: ...
 
 
 class PRCreateUsecase:
